Move McMurdo and South Pole to 'backward'.
This continues in the series of moving entries to 'backward' if
they exist only because of obsolete tz rules about country codes.
This change does not affect any time stamps after 1970.  Unlike
the previous change in this series, this one does affect time
stamps before 1970, but that's OK as pre-1970 time stamps are
out of scope for the tz database.
* antarctica (Antarctica/McMurdo, Antarctica/South_Pole):
Move to 'backward' and link to Pacific/Auckland.
* backward: Move links here from 'antarctica'.
* zone.tab: Coalesce McMurdo and South Pole, and add Scott Base
to the description.