Move obsolescent Montreal entry to 'backward'.
This is like the previous change, except it entails a change to zone.tab.
The America/Montreal entry was created only because we mistakenly
thought that Montreal's and Toronto's post-1970 time stamps differed.
* backward, northamerica (America/Montreal):
Now a 'backward' link to America/Toronto.
* northamerica (Mont): Remove rules; no longer needed.
(America/Thunder_Bay): Refer to Toronto rules, not Mont.
This doesn't change any time stamps.
* zone.tab (America/Montreal): Remove.
(America/Toronto): Say that it also covers most of Quebec.