Fix glitches with '-' and italics in man pages.
Using '\-' to denote ASCII minus has problems in groff output, since
it generates a minus sign in the current font, and if you cut and
paste from the documentation the resulting text won't work.  To avoid
the problem, define a string \*- that expands to ASCII minus in a
constant-width font in groff.  Also, change some instances of \(mi and
\(pl to use plain \- and +, since we no longer need special fonts for
these.  Finally, remove some no-longer-needed instances of "\^" after
italicized words.