Use quote-includes for tzfile.h documentation
The use of angle brackets versus double-quotes for the argument
to the C preprocessor include directive provides some measure of
indication of whether the file being included is from a system
location or a local location.  Since tzfile.h is not installed,
consumers must use it from the source tree, which is by definition
not a system include path.  Use double quotes for the include
directive in the documentation so as to give the reader an
indication that this header is not a system header.