Change Cayman abbreviations to CIST and CIDST
A bill to implement US-style daylight saving time has been
published by the Cayman Islands Gazette, so it looks like our
clock predictions are still on track. The bill gives
Cayman-specific time zone abbreviations, though, so we need to
change tzdata for that.
* NEWS: Document this.
* northamerica (Cayman): New rule.  It is like the US since 2016,
but with different LETTER/S.
(America/Cayman): Use it.