Regularize and resize column-4 comments
This should make these comments more consistent, and make tzselect
easier to use.  Thanks to J William Piggott for reporting the
problem and suggesting a good part of the solution.
* NEWS: Document this.
* iso3166.tab (BQ, MF, SX): Shorten names, so that the tzselect
menu fits on a 24x80 screen after selecting the Americas.
* zone.tab, zone1970.tab: Use a more-standard format for the
column-4 comments.  Reformat so that the tzselect menu fits
on a 24x80 screen even for Canada, Mexico, Russia, and the US.