Port better to MS-Windows
Problems reported by Ian Abbott in:
http://mm.icann.org/pipermail/tz/2016-February/023286.html
http://mm.icann.org/pipermail/tz/2016-February/023287.html
* Makefile (CFLAGS): Add comment about new -D options.
* date.c (environ, optarg, optind, tzname):
* private.h (asctime_r, timezone, daylight, altzone):
* strftime.c (tzname):
* zdump.c (environ, getopt, optarg, optind, tzname):
* zic.c (getopt, link, optarg, optind):
Do not declare if HAVE_POSIX_DECLS, to avoid collisions with
system declarations, which is a problem with MS-Windows
and tzname and the dllimport attribute.
* date.c, zdump.c (tzname): Do not specify size, as POSIX doesn’t.
* private.h (HAVE_POSIX_DECLS): Default to 1.
(ENOTSUP): Default to EINVAL.
* zic.c: If HAVE_DIRECT_H, include direct.h and io.h and
define a mkdir macro, for MS-Windows.
(link, symlink): Set errno to ENOTSUP in the substitutes.
(dolink): Don’t complain merely because link and/or
symlink is not supported.  Be a bit more economical and robust
about checking for directories and existing destinations.
Report errno-related string on link failures.
(itsdir): Work correctly even if a directory has a timestamp that
is out of time_t range, so that stat fails with errno ==
EOVERFLOW.
(writezone): Don’t remove files we can’t stat.