zic is now more careful about symlinks
This avoids some problems in creating nonworking hard links to
symlinks, as well as removing the need for some stat calls.
* NEWS: Mention this.
* zic.c (lstat, S_ISLNK) [!HAVE_SYMLINK]: New fallback macros.
(dolink): Remove destination before linking; this is more consistent
with what is done with fopen, and simplifies the code.  It also lets
us avoid some mkdir calls in some cases.
(itsdir): Document that errno is set when -1 is returned.
Use lstat, not stat.  Avoid tricky code with EOVERFLOW, as it doesn't
work with lstat and probably wasn't worth the trouble anyway.
(writezone): Do not bother statting the destination before removing
it.  The old code catered to ancient platforms that lacked the
'remove' function and where trouble could ensue if a program unlinked
a directory.  We don't need to worry about these platforms any more,
as the code has been successfully using 'remove' since the 1980s.
(mkdirs): Exit on failure, and return true if some directories are made.
All callers changed.  Allow symlinks as well as directories.