Port zdump to C90 + snprintf
Problem reported by Jon Skeet in:
https://mm.icann.org/pipermail/tz/2017-October/025362.html
* NEWS: Mention this.
* zdump.c (my_snprintf): New macro or function.  If a macro, it is
just snprintf.  If a function, it is the same as the old snprintf
static function, with an ATTRIBUTE_FORMAT to pacify modern GCC.
All uses of snprintf changed to use my_snprintf.  This way,
installers don’t need to specify -DHAVE_SNPRINTF if they are using
a pre-C99 compiler with a library that has snprintf.