Use type 0 for timestamps before first transition
I ran into this problem when helping to draft:
https://tools.ietf.org/html/draft-murchison-tzdist-tzif-04
Its section 3.2 gives a complex four-step procedure for
determining the type of timestamps before the first transition,
derived from 2018e localtime.c.  However, I just now checked, and
glibc does only the last two steps.  Furthermore, in practice
nowadays only the last step is needed in almost all zones, because
type 0 is almost always the right type to use; the only exceptions
are the legacy zones EST5EDT, CST6CDT, MST7MDT, PST8PDT, CET, MET,
and EET.  As the tzif-04 section 3.2 procedure is only for
backward compatibility with older TZif files, now is a good time
to clean up by specifying a simple way to deal with time stamps
before the first transition, namely, to use time type 0.
So, this patch changes zic so that the first-used standard time type
(if there is one) is always time type 0.  This is compatible with
glibc and with any other program that uses the heuristic of using
the lowest-numbered standard-time time type for timestamps before
the first transition, since it merely changes zic to output
a standard-time type 0.
* NEWS, tzfile.5: Document this.
* localtime.c (tzparse): Make standard time type 0, if standard
time is in use.  This should help future-proof this program
if we ever get rid of the defaulttype member and just use 0.
* zic.c (swaptypes): New function.
(writezone): Reorder types to make type 0 the default type.
Negate the local variable writetypes into omittypes, so that
strlen can be used on it, and enlarge it by 1 to make room for a
sentinel.  Use memset where appropriate.