Use “timezone“ for timestamp histories
The POSIX standard uses the spelling “timezone”, and often uses it to
mean a TZ setting that determines the past and predicted future of UT
offsets, time zone abbreviations, and tm_isdst flags.  In contrast, in
popular usage by far the most common spelling is “time zone”, and it
typically means a maximal geographic region of clocks that currently
have the same standard time.  As these two quite-different meanings
are often confused, attempt to disambiguate things a bit by
consistently using “timezone” for the former meaning and “time zone”
for the latter.  This affects documentation and a few diagnostic
messages.