Put a bit more info into tzdata.zi version string
Inspired by a remark by Brian Inglis, who suggested REDO in:
https://mm.icann.org/pipermail/tz/2018-July/026664.html
This doesn't do the other things he mentioned, as they are
not supported by our Makefile; if a downstream distributor
implements those other things, the distributor should append
a downstream-specific suffix to the 'version' string.
* Makefile (tzdata.zi): Depend on zishrink.awk.
Pass REDO and dependencies to zishrink.awk, instead of
just BACKWARD and PACKRATDATA.
* NEWS: Mention this.
* zishrink.awk: Report any changes to list of input files, not
merely changes to BACKWARD or PACKRATDATA.  Output nondefault REDO
as well; although REDO doesn't affect tzdata.zi it does affect
whether leap seconds are intended to be in the generated TZif
files, and this might be useful.

2018-07-18  Paul Eggert  <eggert@cs.ucla.edu>

* Makefile (AWK): Mention mawk problems.