One line per metadatum in tzdata.zi
With the new metadata, version lines could be long.  E.g., 'make
DATAFORM=rearguard REDO=posix_only PACKRATDATA=backzone BACKWARD='
could generate a tzdata.zi line like this:
  # version 2018e-57-gba7f088,dataform=rearguard,redo=posix_only,ddeps=(backzone,!backward)
which is quite a mouthful and is stretching the definition of
"version".  To avoid the problem, output a separate comment line
for each metadata component instead, like this:
  # version 2018e-57-gba7f088
  # dataform rearguard
  # redo posix_only
  # ddeps backzone !backward
This is easier to read and to parse and should scale better
if we add more metadata later.
* NEWS: Mention this.
* zishrink.awk: Output a separate line for each metadatum.

2018-07-21  Paul Eggert  <eggert@cs.ucla.edu>

Use space, not comma, in tzdata.zi "# version"
* zishrink.awk: Use space instead of comma to separate original
version string from auxiliary configuration information.  This
saves two bytes and (more important) is easier for humans to read.