Fix appended dummy transition in non-POSIX zones
* zic.c (outzone): When adding a redundant transition to indicate
definite knowledge of a lack of transitions in a zone whose
predicted future cannot be represented in a POSIX TZ string, the
transition’s type should be the same as the previous transition;
it should not be the maximum type value.  The distinction matters
in the upcoming Africa/Casablanca zone; without this patch, zic
would generate an incorrect year-2440 transition for the zone,
where 2440 = 2037 + 400 + 3.