zic: fix transitions created with -r
It’s nicer (though not required by RFC 8536) that when the start
of a TZif file is truncated, the default time type is that of
timestamps just before the first transition.
Problem reported by Tim Parenti in:
https://mm.icann.org/pipermail/tz/2019-March/027779.html
Also, implement the RFC 8536 requirement that when the end of a
TZif file is truncated, the file should have a 64-bit transition
designating the truncation boundary.
* zic.c (puttzcod64): Remove, replacing with ...
(puttzcodepass): ... this new function, which has a new arg PASS.
All uses changed.
(limitrange): Use a simpler approach to discarding transitions
out of range, so that the caller can restore them if needed.
Fix off-by-one error when discarding too-high transitions.
(writezone): Use proper type for first transition, if the
file was truncated at the start.  Create a last transition if
the file was truncated at the end.