Assume zic 2017c or later when shrinking
The tzdata.zi file was first released in 2017c, and deliberately
avoided zic features that were also new to that release, to avoid
problems with people mistakenly running older zic when generating
this new file.  After a year’s wait it is reasonable to assume
that anybody using this new file is also using a new-enough zic.
This saves about 900 bytes in tzdata.zi.
* NEWS: Mention this.
* zishrink.awk (gen_rule_name, prehash_rule_names)
(process_input_line, output_saved_lines):
For clarity, use the convention that parameters are on the first
line of function definition, and local vars are on later lines.
(process_input_line): Assume zic 2017c or later.
Make line types more explicit.