Update URLs
* northamerica, theory.html, tz-art.html, tz-how-to.html, tz-link.html:
Update URLs.  A lot of this is replacing "http:" by "https:" when
either will do; most of the rest is updating URLs that moved.
The substantive changes are upgrading from ISO 8601:2004 to ISO
8601-1:2019, adding a C++20 link, removing URLs for DJGPP and
Walltime-js (now obsolete), removing URLs maintained by the
National Physical Laboratory and astrologyschool.com (no longer
existing), and saying that draft-murchison-tzdist-geolocate is
expired.