Omit a few artifacts from zic output
* NEWS: Mention this.
* zic.c (writezone): When optimizing, discard a transition if it
is marked as mergable and if it changes neither the UT offset, nor
the isdst flag, nor the time zone abbreviation.  This removes a
few useless transitions and types from the output when they are
merely artifacts of how the input was written.  For example,
without this fix Europe/London generates a useless transition
at the start of 1996, because the line "0:00 GB-Eire %s 1996"
has an UNTIL field that specifies 1996-01-01 00:00:00 *local time*,
unlike the previous transition specified at 1995-10-22 01:00:00 *UTC*,
which means addtype will incorrectly consider the 1996 transition
as being significant (as the two time types differ in their
ttisgmt value even though they do not differ in UT offset, isdst
flag, or time zone abbreviation).