Fix 'zic -b slim' bug with TZ='CET'
Without this fix, 'zic -b slim' can mishandle transitions in the
few obsolescent zones like CET that do not start with local mean time.
For example, for CET it incorrectly optimizes away the first and
third transitions, and treats the second transition as one to
standard time instead of to DST.
* zic.c (swaptypes): Remove.  All uses removed.
(writezone): Swap indices rather than contents of the old type 0
and the new (default) type 0 array entries.