More clarification/explanation from RFC 8536
This just changes documentation: it does not change code or data.
Mostly, it just adopts more terminology from RFC 8536.
It also contains an example of what the std/wall and UT/local
indicators are for.
* newtzset.3, theory.html, tz-how-to.html, tzfile.5, tzfile.h, zic.8:
In documentation and comments, make it clearer that local time
equals wall clock time.
* tzfile.5, tzfile.h: In documentation and comments, say that when
isut is 1, isstd must also be 1.
* tzfile.5: Specify that bytes are 8 bits, that binary integers
use two's complement, and that booleans use 0 and 1.
Say that UT offsets are never equal to -2**31, and that they
realistically are in the range [-89999, 93599].  Say that
leap second counts are signed.  Say why std/wall and UT/local
indicators are present, and give AKST9AKDT as an example.
Do not say that the first standard-time ttinfo is used by
localtime, as the actual heuristic is far more complicated;
instead, document what RFC 8536 says as what's normally used.
* zic.c (struct rule): Adjust comments.