Deprecate ‘zic -p POSIXRULES’
zic’s -p option was intended as a transition from historical
System V code that treated TZ="XXXnYYY" as meaning US
daylight-saving rules in a time zone n hours west of UT,
with XXX abbreviating standard time and YYY abbreviating DST.
zic -p allows the tzdata installer to specify (say)
Europe/Brussels's rules instead of US rules.  This behavior
is not well documented and often fails in practice; for example it
does not work with current glibc for contemporary timestamps, and
it does not work in tzdb itself for timestamps after 2037.
So, document it as being obsolete, with the intent that it
will be removed in a future version.  This change does not
affect behavior of the default installation.
* Makefile, NEWS, tzfile.5, zic.8:
Adjust comments and documentation accordingly.
* Makefile (install): Add support for empty POSIXRULES, which now
means not to use zic -p when installing.