Allow definition of vars the system lacks
For example, FreeBSD lacks ‘altzone’ but we may want to
support it in the replacement library.
* Makefile, NEWS: Document change.
* localtime.c (tzname, timezone, daylight, altzone): Define only
if necessary: either the corresponding macro is 2, or it is 1 and
TZ_TIME_T is set so we are implementing the features under a
different name.
* private.h (ALTZONE): Default to 1 or 0.
All uses now changed to ‘#if’ instead of ‘#ifdef’.
(tzname, timezone, daylight, altzone):
Declare only if necessary: either the corresponding macro is 2,
or it is 1 and TZ_TIME_T is set (so we are implementing the
features under a different name) or !HAVE_POSIX_DECLS (so the
system headers lack the decls).