Update Hong Kong section
No one ever calls Hong Kong as Xianggang, in the past or currently. At one time Hong Kong was labeled as such years ago on Google Maps, which was likely the result of automatic romanization of Mandarin. The Pinyin romanization scheme of the Mandarin pronunciation of 香港 is not officially recognized or even mentioned in official documents.

https://www.basiclaw.gov.hk/en/basiclawtext/chapter_1.html

https://www.elegislation.gov.hk/hk/cap5