Change Europe/Kiev to Europe/Kyiv to use correct spelling
Disclaimer: Language question is tough for Ukraine, and I have no
            intention to speculate on the topic. The PR is about
            concrete renaming Kiev to Kyiv according to national
            and worldwide version of spelling of the capital of
            Ukraine.

I would agree with comment of @eggert, that transliteration of "Київ"
to Kiev is wrong and for some reasons was popular.
However, I do believe we should use the real city name instead
of what was "popular" to avoid any further confusion. Popularity
is a volatile thing, and can't be used in systems which supposed to
be consistent. That will generate a lot of whys and can lead to
misunderstanding.

Described example about Praha and Prague doesn't make much sense.
I think governments of both countries highlight that. The
Government of the Czech Republic uses "Prague" (1) when
the Cabinet of Ministers of Ukraine uses "Kyiv" (2). What I'm
trying to say, is that even if Prague is written incorrectly,
it accepted by society.
Nowadays, the majority of map services, such as google/maps,
openstreetmap, mapbox use Kyiv as well when you search for
the capital of Ukraine.

Moreover, "Kyiv" is used by other governments and press (3,4,5,6).
Not all of them, of course, but I think that may change anytime
soon.

Links:
1. https://www.vlada.cz/en/urad-vlady/adresa-uradu-vlady/contact-address-50831/
2. https://www.kmu.gov.ua/en/kontakti
3. https://www.ft.com/stream/8e326860-810f-3f96-8b08-8db10ef58906
4. https://www.nytimes.com/search?query=kyiv
5. https://www.gov.uk/world/organisations/british-embassy-kyiv
6. https://www.rferl.org/a/us-to-change-international-database-spelling-to-kyiv-not-kiev/29997415.html