Port to downstream HP-UX-style ‘make’
HP-UX ‘make’ considers a target to be out-of-date if its timestamp
equals that of a dependency, and POSIX allows this ‘make’ behavior.
So, when putting timestamps on files in tarballs, add 1 to a
file’s timestamp if it depends on some other file with the same
timestamp.  That way, when someone unpackages a tarball on an
HP-UX host, its ‘make’ will avoid rebuilding files like ‘version’
that are up-to-date in the tarball.
* Makefile (SET_TIMESTAMP_N, SET_TIMESTAMP, SET_TIMESTAMP_DEP):
New macros.
(set-timestamps.out, set-tzs-timestamp.out)
(tzdata$(VERSION)-rearguard.tar.gz, tzdb-$(VERSION).tar.lz):
Use them instead of invoking ‘touch’ by hand.
(set-timestamps.out): Set the timestamp of tzdata.zi after that
of version, since the former depends on the latter.