Port 'make rearguard_tarballs' to Solaris 10
* Makefile (tarballs, etc.): Export AWK to submake, for Solaris 10
which needs something like 'make AWK=nawk rearguard_tarballs'.
(.KEEP_STATE): Remove.  .KEEP_STATE is a Solaris 'make' feature
that breaks 'make rearguard_tarballs' because it causes submakes
to mistakenly consider files to be out-of-date.  As .KEEP_STATE
is not specified by POSIX and is ignored by GNU 'make' it should
not be needed.