Avoid backward links in zone.tab
Recent changes caused zone.tab to refer to links defined in the
'backward' file, which means it no longer works in the now-rare case
where 'backward' is not used.  Fix this by changing zone.tab to refer
to the primary name of each zone.
* checktab.awk: Adjust to check the new setup.
* zone.tab: Replace backward links in the 3rd column with the primary
zone names.  This lets us fix a longstanding problem with Vietnam, as
zone.tab can now have a separate entry for north Vietnam's different
time zone history in the 1970s (this was already true of zone1970.tab).