Replace Pacific/Enderbury with Pacific/Kanton
When I added Pacific/Enderbury in 1993, I thought it was the
most-populous location in its timezone.  I was wrong - it has been
uninhabited since 1942.  I discovered my mistake this week when
doing the recent timezone-coalescing changes.  It shouldn’t be
surprising that this problem has not been reported, as only two
dozen people live in the timezone, they’re rotated in and out, and
they likely don’t have a good Internet connection.  Leaving the
name alone wouldn’t work because Pacific/Enderbury should be at
-00 now, our convention for uninhabited locations.  So, replace
the name with Pacific/Kanton.  By strange coincidence, Kanton has
been in the news recently, for reasons similar to why Enderbury
was in the news in the 1930s; see:
https://www.thedrive.com/the-war-zone/40471/china-wants-to-revive-a-strategically-located-military-airfield-deep-in-pacific-report
* NEWS: Mention this.
* australasia, zone.tab, zone1970.tab (Pacific/Kanton):
New zone, replacing ...
* backward (Pacific/Enderbury): New backward-compatibility link
* backzone: Copy of old data, with fixes for uninhabited times,
which includes the current time.