Fix ambiguous leapsecs by rolling up to a minute
This fixes a longstanding ambiguity in the output of localtime in
oddball timezones where a positive leap second occurs when the UTC
offset is not a multiple of 60 seconds.  See NEWS for details.
* NEWS: Describe fix.
* localtime.c (timesub): Use "??:??:60" at the end of the
localtime minute containing the leap second, even when the second
marked "??:??:60" is not itself a leap second.  This avoids
ambiguity in localtime output, at the cost of having the localtime
leap second occur up to a minute away from the UTC leap second
when the UTC offset is not a multiple of 60 seconds.
* tzfile.5: Document the fixed and unfixed behaviors.