Don't mistruncate TZif files at leap seconds
Without this fix, zic mishandles the command
‘zic -L leapseconds -r @1706745601 etcetera’ when ‘leapseconds’
is generated from the test data file mentioned in:
https://mm.icann.org/pipermail/tz/2021-September/030385.html
The resulting, incorrect TZif file causes
‘TZ=Etc/GMT date -r 1706745601’ to output
"Thu Feb  1 00:00:01 GMT 2024" instead of the correct
"Thu Feb  1 00:00:00 GMT 2024", because localtime misinterprets
the leap second entry at 1706745601 to be a positive instead of a
negative leap second.
* NEWS: Mention the fix.
* zic.c (limitrange): When omitting leading leap seconds, do not
output a leap second table that starts with a positive leap second
with a nonpositive correction (or with a negative leap second with
a positive correction) as that will confuse TZif readers into
thinking the positive leap second is negative or vice versa.
Instead, keep otherwise-unnecessary leading leap seconds to
avoid the confusion.