Be more generous when dealing with v1 data
* NEWS: Mention this.
* localtime.c (tzloadbody): In version 2 and later files, do not
attempt to validate the version 1 header and data block; instead,
ignore them except for the purpose of skipping over them, as per a
recommendation in RFC 8536 section 4.
Also, when the TZif file is version 1 and we break out near the
end of the loop, do so only after altering nread and memmoving any
later data; this avoids having later code attempt to parse the
version 1 header as if it were a TZ string.