Simplify field memory allocation in zic
* zic.c (MAX_FIELDS): New macro.
(infile): Allocate the fields array on the stack, since it’s small.
That way there’s no need to free it explicitly later.
(getfields): New args ARRAY and ARRAYELTS, so that there’s no need
to allocate an array.  Return number of fields found.  Report an
error if there are too many fields.