Fix bug uncovered by recent change to Iran history
Without this patch, the obvious fix for Iran, which can be seen by
using the next patch after this but commenting out the line that works
around this bug, results in the bogus diagnostic "can't determine time
zone abbreviation to use just after until time".
* zic.c (disable_percent_s): New static constant.
(doabbr): Refuse to expand %s to %s if LETTERS is this constant.
(outzone): Even if the rule is after the UNTIL, use it to
specify the abbreviation if the abbreviation is not known yet.
Also, if FORMAT column doesn't use %s, the rule need not be
consulted to determine the abbreviation.