Check that ziguard substitutions revert
Add a check that substitutions to vanguard or rearguard form can
be reverted.  This supports possible future uses when continental
files are in vanguard form and you want rearguard (or vice versa).
* Makefile (check): Depend on ...
(check_ziguard): ... this new target.
* ziguard.awk: Fix some in_comment typos, as these column
numbers do not depend on whether we’re in a comment.
Use sub(X,Y) instead of $I = Y to preserve whitespace.
Fix Japan and Morocco substitutions so that they are reversible.