Port .txt generation to macOS
Be able to build man pages on macOS, which has mandoc not nroff.
Although mandoc has some minor glitches compared to nroff,
it’s good enough.  This follows up on a suggestion by Guy Harris in:
https://mm.icann.org/pipermail/tz/2022-August/031786.html
* time2posix.3: Insert \& to convince groff to not double-space
afterwards, so that it doesn’t disagree with mandoc.
* tzselect.8, zic.8: Use \& instead of \" because mandoc
mishandles the latter.  Don’t use a 5-deep nested .if on one line,
as mandoc mishandles it.
* workman.sh: If nroff and perl are not both available,
then use mandoc and col if they are both available,
else report an error.