Fix bug with 'zic -d /a/b/c' if /a is unwriteable
If the directory /a/b already exists, it is not an error if /a is
unwritable when you invoke zic -d /a/b/c.  Problem found on
Solaris 10, where mkdir("/a", ...) can fail with EACCES in this
situation, which POSIX allows.
* NEWS: Mention the fix.
* private.h (ELOOP) [!ELOOP]: Define to EINVAL.
* zic.c (mkdirs): Do not fail merely because mkdir fails
with errno == EACCES, as the parent directory may already
exist but be unwritable.  For example