Improve tzselect behavior for Turkey etc.
Fix ‘tzselect’ so that you can choose Istanbul regardless of
whether you choose Asia or Europe at the top level, and
similarly for other Zones that straddle continents/oceans.
Inspired by a problem reported by Peter Krefting in:
https://mm.icann.org/pipermail/tz/2022-August/031795.html
* tzselect.ksh (quoted_continents, countries): Parse new #@ lines
and use them to deduce continents and countries better.
* zone1970.tab: Add #@ comments for Zones whose continent
or ocean are not easily deducible.