Treat 1931 changes in Mexico as DST
Mexico’s Chamber of Deputies was supposed to vote today on
legislation abolishing DST in most of Mexico.  Although the vote
was delayed due to Internet problems, news coverage of the
non-event led me to review our sources for DST in Mexico, and I
found that tzdb considered the transitions in 1931 to be two
changes to standard time, whereas it is considered to be DST in
today’s news coverage, as well as in our main source, the history
of Mexican local time published by the Investigation and Analysis
Service of the Mexican Library of Congress.  Fix this by
considering the 1931 changes to be DST, not standard time.
* NEWS: Mention this.
* northamerica (Mexico): Add DST rules for 1931.
(America/Mexico_City, America/Ojinaga, America/Chihuahua)
(America/Hermosillo, America/Mazatlan, America/Bahia_Banderas):
Use them.